requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
alembic==1.14.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.2.0
auth0-python==4.7.2
backoff==2.2.1
bcembedding==0.1.5
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.21
click==8.1.7
cohere==5.13.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
crewai==0.83.0
crewai-tools==0.14.0
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.0.1
decorator==5.1.1
deprecated==1.2.15
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
docker==7.1.0
docstring-parser==0.16
docx2txt==0.8
durationpy==0.9
embedchain==0.1.125
entrypoints==0.4
et-xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.8.0.post1
faker==37.1.0
fastapi==0.115.5
fastavro==1.9.7
favicon==0.7.0
filelock==3.13.1
flask==3.1.2
flatbuffers==24.3.25
fonttools==4.55.3
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
gitpython==3.1.43
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-aiplatform==1.73.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googletrans==4.0.2
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
grpcio-tools==1.68.1
gtts==2.5.4
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
htbuilder==0.9.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.25.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.0.0
instructor==1.7.0
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json-repair==0.30.2
jsonpatch==1.33
jsonpickle==4.0.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.8
kubernetes==31.0.0
lancedb==0.16.0
langchain==0.3.9
langchain-cohere==0.3.3
langchain-community==0.3.9
langchain-core==0.3.21
langchain-experimental==0.3.3
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langsmith==0.1.147
litellm==1.53.2
lxml==5.4.0
mako==1.3.6
markdown==3.8
markdown-it-py==3.0.0
markdownlit==0.0.7
markupsafe==2.1.5
marshmallow==3.22.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.34
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
mysqlclient==2.2.4
narwhals==1.37.0
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.3
onnxruntime==1.20.1
openai==1.56.0
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.12
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pdfminer-six==20231228
pdfplumber==0.11.4
pillow==10.2.0
pip==24.2
plotly==6.0.1
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.4
prometheus-client==0.21.1
prompt-toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.2
pure-eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyaudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.6.1
pydeck==0.9.1
pygments==2.18.0
pyjwt==2.10.1
pylance==0.19.2
pymdown-extensions==10.15
pymysql==1.1.1
pyparsing==3.2.1
pypdf==5.0.1
pypdfium2==4.30.0
pypika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.389
pysbd==0.3.4
pysocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
pywin32==308
pyyaml==6.0.2
qdrant-client==1.12.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.27.1
sentence-transformers==3.0.1
setuptools==75.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
speechrecognition==3.11.0
sqlalchemy==2.0.35
st-annotated-text==4.0.2
st-theme==1.2.3
stack-data==0.6.3
starlette==0.41.3
streamlit==1.39.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-chat==0.1.1
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-mic-recorder==0.0.8
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.2.1
tomli-w==1.1.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.36.2
trio==0.27.0
trio-websocket==0.11.1
typer==0.14.0
types-requests==2.32.0.20241016
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
uv==0.5.5
uvicorn==0.32.1
validators==0.34.0
watchdog==5.0.3
watchfiles==1.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.13.1
zipp==3.21.0